rename schemas folder back to models to avoid merge issues
