+++ api/controllers/degree.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { DegreeModel } from '../schemas/degree';
+import { DegreeModel } from '../models/degree';
 
 export const degreeSearch = async (req: Request, res: Response) => {
   DegreeModel.find(req.query, {}, { strict: false }, (error, result) => {
