+++ api/models/degree.ts
